# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Practica Parcial 1'","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'"}],"paths":{"/weather/subscribe":{"post":{"tags":["weather-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/weather/locations/{client_id}/{client_secret}":{"get":{"tags":["weather-controller"],"operationId":"getLocations","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"client_secret","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationClientResponse"}}}}}}}},"/weather/locations/location/{id}/{datetime}/{client_id}/{client_secret}":{"get":{"tags":["weather-controller"],"operationId":"getLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"client_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"client_secret","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"datetime","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocationInformation"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"UserDTO":{"required":["email","temperature_unit"],"type":"object","properties":{"email":{"type":"string"},"temperature_unit":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"secret":{"type":"integer","format":"int64"},"temperature_unit":{"type":"string"}}},"LocationClientResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"AirQuality":{"type":"object","properties":{"index":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"Cloudiness":{"type":"object","properties":{"index":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"Location":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"LocationInformation":{"type":"object","properties":{"location":{"$ref":"#/components/schemas/Location"},"temperature":{"$ref":"#/components/schemas/Temperature"},"wind":{"type":"string"},"air_quality":{"$ref":"#/components/schemas/AirQuality"},"cloudiness":{"$ref":"#/components/schemas/Cloudiness"}}},"Temperature":{"type":"object","properties":{"value":{"type":"number","format":"double"},"unit":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}