# src/main/java/ar/edu/utn/frc/tup/lciii/models/AirQuality.java
package ar.edu.utn.frc.tup.lciii.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AirQuality {
    private Integer index;
    private String description;
}