# src/main/java/ar/edu/utn/frc/tup/lciii/models/Cloudiness.java
package ar.edu.utn.frc.tup.lciii.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Cloudiness {
    private Integer index;
    private String description;
}