# src/main/java/ar/edu/utn/frc/tup/lciii/models/Temperature.java
package ar.edu.utn.frc.tup.lciii.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Temperature {
    private Double value;
    private String unit;
}